+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/ProfileFormOverlay/ProfileFormOverlay.js
@@ -6,7 +6,7 @@ import Loader from '../../components/Loader/Loader';
 import Overlay from '../../components/Overlay';
 import Form from '../../containers/Form';
 import ResourceRequester from '../../services/ResourceRequester';
-import userStore from '../../stores/UserStore';
+import userStore from '../../stores/userStore';
 import {translate} from '../../utils/Translator';
 import MemoryFormStore from '../Form/stores/MemoryFormStore';
 import metadataStore from '../Form/stores/metadataStore';
