+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/ProfileFormOverlay/tests/ProfileFormOverlay.test.js
@@ -3,7 +3,7 @@ import {shallow, mount} from 'enzyme/build';
 import mockReact from 'react';
 import Overlay from '../../../components/Overlay';
 import ResourceRequester from '../../../services/ResourceRequester';
-import userStore from '../../../stores/UserStore';
+import userStore from '../../../stores/userStore';
 import Form from '../../Form';
 import ProfileFormOverlay from '../ProfileFormOverlay';
 const React = mockReact;
@@ -25,7 +25,7 @@ jest.mock('sulu-admin-bundle/containers/Form/stores/MemoryFormStore', () => clas
         }
 });
 
-jest.mock('sulu-admin-bundle/stores/UserStore', () => ({
+jest.mock('sulu-admin-bundle/stores/userStore', () => ({
     setFullName: jest.fn(),
 }));
