+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/index.js
@@ -3,7 +3,7 @@ import localizationStore from './LocalizationStore';
 import type {Localization} from './LocalizationStore/types';
 import MultiSelectionStore from './MultiSelectionStore';
 import ResourceStore from './ResourceStore';
-import userStore from './UserStore';
+import userStore from './userStore';
 
 export {
     MultiSelectionStore,
