+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/userStore/types.js
