+++ src/Sulu/Bundle/AdminBundle/Resources/js/stores/userStore/userStore.js
