+++ src/Sulu/Bundle/AdminBundle/Resources/js/views/List/tests/List.test.js
@@ -19,7 +19,7 @@ jest.mock('../../../containers/List/stores/metadataStore', () => ({
     getSchema: jest.fn().mockReturnValue({}),
 }));
 
-jest.mock('../../../stores/UserStore', () => ({
+jest.mock('../../../stores/userStore', () => ({
     setPersistentSetting: jest.fn(),
     getPersistentSetting: jest.fn(),
 }));
